Models now specified as functions
Refactored this so that models must be a list of functions. The default odds ratio and hazard ratio models are now specified as closures, allowing for a simple API that includes function parameters in the call. This also allows for user-specified models to be included in the same call as these defaults, so there is no distinction in contingency table between custom models and the defaults.